+++ Rebus/Transport/SqlServer/DisabledTimeoutManager.cs
@@ -7,7 +7,7 @@ using Rebus.Timeouts;
 
 #pragma warning disable 1998
 
-namespace Rebus.AzureStorageQueues
+namespace Rebus.Transport.SqlServer
 {
     class DisabledTimeoutManager : ITimeoutManager
     {
@@ -17,8 +17,8 @@ namespace Rebus.AzureStorageQueues
 
             var message = string.Format("Received message with ID {0} which is supposed to be deferred until {1} -" +
                                         " this is a problem, because the internal handling of deferred messages is" +
-                                        " disabled when using Azure Storage Queues as the transport layer in, which" +
-                                        " case the native support for a specific initial visibility delay is used...",
+                                        " disabled when using SQL Server as the transport layer in, which" +
+                                        " case the native support for a specific visibility time is used...",
                 messageIdToPrint, approximateDueTime);
 
             throw new InvalidOperationException(message);